requirements.txt
absl-py==2.3.1
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aioice==0.10.1
aiortc==1.14.0
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==3.7.1
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
astrapy==2.2.1
asttokens==3.0.1
astunparse==1.6.3
async-lru==2.0.5
attrs==25.4.0
av==16.0.1
azure-ai-agents==1.2.0b6
azure-ai-projects==1.1.0b4
azure-core==1.36.0
azure-identity==1.25.1
azure-storage-blob==12.27.1
babel==2.17.0
beautifulsoup4==4.14.2
black==26.1.0
bleach==6.3.0
blinker==1.9.0
cachetools==6.2.6
cassandra-driver==3.29.3
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
cloudevents==1.12.0
comm==0.2.3
contourpy==1.3.3
coverage==7.13.1
cryptography==46.0.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
DoubleMetaphone==1.2
executing==2.2.1
faiss-cpu==1.13.2
fastapi==0.104.1
fastjsonschema==2.21.2
filelock==3.20.0
flake8==7.3.0
Flask==3.1.3
flatbuffers==25.9.23
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.10.0
gast==0.6.0
geomet==1.1.0
gitdb==4.0.12
GitPython==3.1.46
google-crc32c==1.7.1
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
h5py==3.15.1
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
ifaddr==0.2.0
importlib_metadata==8.7.0
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isodate==0.7.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
joblib==1.3.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
keras==3.12.0
kiwisolver==1.4.9
lark==1.3.1
lazy-object-proxy==1.12.0
Levenshtein==0.27.3
libclang==18.1.1
librt==0.7.8
llvmlite==0.45.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.4
ml_dtypes==0.5.4
more-itertools==10.8.0
mpmath==1.3.0
msal==1.34.0
msal-extensions==1.3.1
multidict==6.7.0
mypy==1.19.1
mypy_extensions==1.1.0
mysql-connector-python==9.6.0
namex==0.1.0
narwhals==2.16.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
neo4j==6.1.0
nest-asyncio==1.6.0
networkx==3.6
notebook==7.5.0
notebook_shim==0.2.4
numba==0.62.1
numpy==1.25.2
openai==1.109.1
openapi-core==0.19.5
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opt_einsum==3.4.0
optree==0.18.0
overrides==7.7.0
packaging==25.0
pandas==2.1.3
pandocfilters==1.5.1
parse==1.20.2
parso==0.8.5
pathable==0.4.4
pathspec==1.0.3
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
plotly==6.5.2
pluggy==1.6.0
prance==25.4.8.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.0
pybars4==0.9.13
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.5.2
pydantic-settings==2.11.0
pydantic_core==2.14.5
pydeck==0.9.1
pyee==13.0.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pylibsrtp==1.0.0
PyMeta3==0.5.1
pymongo==4.16.0
pynndescent==0.5.13
pyOpenSSL==25.3.0
pyparsing==3.2.5
pytest==9.0.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.6
pytokens==0.4.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.14.5
referencing==0.36.2
regex==2026.4.4
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.28.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
ruff==0.14.14
scikit-learn==1.3.2
scipy==1.16.3
seaborn==0.13.2
semantic-kernel==1.37.1
Send2Trash==1.8.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.23
stack-data==0.6.3
starlette==0.27.0
streamlit==1.54.0
sympy==1.14.0
tenacity==9.1.4
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
torch==2.9.1
torchvision==0.24.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
umap-learn==0.5.9.post2
Unidecode==1.4.0
uri-template==1.3.0
urllib3==2.5.0
uuid6==2025.0.1
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.1
widgetsnbextension==4.0.15
wrapt==2.0.1
xgboost==3.2.0
yarl==1.22.0
zipp==3.23.0
